Separate internals helpers
Initial commit to separate internals helpers from the main syscall helper. more cleanup in that regard will be added later.
